Fixed Collision Resolution
Momentum now works in the direction that the objects colliding was
moving, rather than in the direction of the normal of the edge they
collide on.

Finding normals of edges of objects is now more robust. The normal is
taken of the line which intersects the line formed by connecting the avg
coordinate penetrating the object with the object's centre of gravity.

As before, the direction normals are pointing in is checked by comparing
in which direction it points to whether the centre of gravity of the
object it relates to is above or below the edge of which the normal is
taken.

Hopefully this is the final commit regarding collision resolution...